Create Phase2&3.py
In phase 2, further analysis was conducted to understand what the machine learning model had learned. The goal was to design an algorithm to determine the contribution of each atom or functional group to the prediction results. The network was constructed to enable the operator to train the model and examine the composite gradients across the entire network. I then optimize the existing ligands to design more potent and selective compounds in Phase 3.